+++ fpdf/fpdf.py
@@ -8,7 +8,15 @@
 # * Maintainer:  David Alexander (daveankin@gmail.com) et al since 2017 est. *
 # * Maintainer:  Lucas Cimon et al since 2021 est.                           *
 # ****************************************************************************
-import hashlib, io, logging, math, os, re, sys, types, warnings
+import hashlib
+import io
+import logging
+import math
+import os
+import re
+import sys
+import types
+import warnings
 from collections import defaultdict
 from contextlib import contextmanager
 from datetime import datetime, timezone
@@ -20,8 +28,8 @@ from pathlib import Path
 from typing import Callable, Dict, Iterator, NamedTuple, Optional, Union
 
 try:
-    from endesive import signer
     from cryptography.hazmat.primitives.serialization import pkcs12
+    from endesive import signer
 except ImportError:
     pkcs12, signer = None, None
 
@@ -39,26 +47,26 @@ except ImportError:
 
 from .actions import URIAction
 from .annotations import (
+    DEFAULT_ANNOT_FLAGS,
     AnnotationDict,
     PDFAnnotation,
     PDFEmbeddedFile,
-    DEFAULT_ANNOT_FLAGS,
 )
 from .bidi import BidiParagraph, auto_detect_base_direction
 from .deprecation import (
-    support_deprecated_txt_arg,
-    get_stack_level,
     WarnOnDeprecatedModuleAttributes,
+    get_stack_level,
+    support_deprecated_txt_arg,
 )
 from .drawing import (
-    convert_to_device_color,
     DeviceRGB,
+    DrawingContext,
     GraphicsStateDictRegistry,
     GraphicsStyle,
-    DrawingContext,
     PaintedPath,
     Point,
     Transform,
+    convert_to_device_color,
 )
 from .encryption import StandardSecurityHandler
 from .enums import (
@@ -72,6 +80,7 @@ from .enums import (
     EncryptionMethod,
     FileAttachmentAnnotationName,
     MethodReturnValue,
+    PageLabelStyle,
     PageLayout,
     PageMode,
     PathPaintRule,
@@ -109,9 +118,10 @@ from .line_break import (
 )
 from .outline import OutlineSection
 from .output import (
+    ZOOM_CONFIGS,
     OutputProducer,
     PDFPage,
-    ZOOM_CONFIGS,
+    PDFPageLabel,
     stream_content_for_raster_image,
 )
 from .recorder import FPDFRecorder
@@ -121,6 +131,7 @@ from .svg import Percent, SVGObject
 from .syntax import DestinationXYZ, PDFArray, PDFDate
 from .table import Table, draw_box_borders
 from .text_region import TextRegionMixin, TextColumns
+from .transitions import Transition
 from .unicode_script import UnicodeScript, get_unicode_script
 from .util import get_scale_factor, Padding
 
@@ -281,8 +292,13 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
         self.oversized_images = None
         self.oversized_images_ratio = 2  # number of pixels per UserSpace point
         self.struct_builder = StructureTreeBuilder()
-        self._toc_placeholder = None  # optional ToCPlaceholder instance
-        self._outline = []  # list of OutlineSection
+        self.toc_placeholder = None  # optional ToCPlaceholder instance
+        self._outline: list[OutlineSection] = []  # list of OutlineSection
+        # flag set true while rendering the table of contents
+        self.in_toc_rendering = False
+        # allow page insertion when writing the table of contents
+        self._toc_allow_page_insertion = False
+        self._toc_inserted_pages = 0  # number of pages inserted
         self._sign_key = None
         self.title = None
         self.section_title_styles = {}  # level -> TextStyle
@@ -844,8 +860,35 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
         """
         self.str_alias_nb_pages = alias
 
+    def set_page_label(
+        self,
+        label_style: Union[str, PageLabelStyle] = None,
+        label_prefix: str = None,
+        label_start: int = None,
+    ):
+        current_page_label = (
+            None if self.page == 1 else self.pages[self.page - 1].get_page_label()
+        )
+        new_page_label = None
+        if label_style or label_prefix or label_start:
+            label_style = (
+                PageLabelStyle.coerce(label_style, case_sensitive=True)
+                if label_style
+                else None
+            )
+            new_page_label = PDFPageLabel(label_style, label_prefix, label_start)
+        self.pages[self.page].set_page_label(current_page_label, new_page_label)
+
     def add_page(
-        self, orientation="", format="", same=False, duration=0, transition=None
+        self,
+        orientation: str = "",
+        format: str = "",
+        same: bool = False,
+        duration: float = 0,
+        transition: Transition = None,
+        label_style: Union[str, PageLabelStyle] = None,
+        label_prefix: str = None,
+        label_start: int = None,
     ):
         """
         Adds a new page to the document.
@@ -869,6 +912,15 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
                 from another page to the given page during a presentation.
                 Can be configured globally through the `.page_transition` FPDF property.
                 As of june 2021, onored by Adobe Acrobat reader, but ignored by Sumatra PDF reader.
+            label_style (str or PageLabelStyle): Defines the numbering style for the numeric portion of each
+                page label. Possible values are:
+                - "D": Decimal Arabic numerals.
+                - "R": Uppercase Roman numerals.
+                - "r": Lowercase Roman numerals.
+                - "A": Uppercase letters (A to Z for the first 26 pages, followed by AA to ZZ, etc.).
+                - "a": Lowercase letters (a to z for the first 26 pages, followed by aa to zz, etc.).
+            label_prefix (str): Prefix string applied to the page label, preceding the numeric portion.
+            label_start (int): Starting number for the first page of a page label range.
         """
         if self.buffer:
             raise FPDFException(
@@ -885,11 +937,28 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
         char_spacing = self.char_spacing
         dash_pattern = self.dash_pattern
 
-        if self.page > 0:
+        in_toc_extra_page = (
+            self.in_toc_rendering
+            and self._toc_allow_page_insertion
+            and self.page > self.toc_placeholder.start_page
+        )
+        if self.page > 0 and (not self.in_toc_rendering or in_toc_extra_page):
             # Page footer
             self.in_footer = True
             self.footer()
             self.in_footer = False
+
+        current_page_label = (
+            None if self.page == 0 else self.pages[self.page].get_page_label()
+        )
+        new_page_label = None
+        if label_style or label_prefix or label_start:
+            label_style = (
+                PageLabelStyle.coerce(label_style, case_sensitive=True)
+                if label_style
+                else None
+            )
+            new_page_label = PDFPageLabel(label_style, label_prefix, label_start)
 
         # Start new page
         self._beginpage(
@@ -900,6 +969,8 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
             transition or self.page_transition,
             new_page=not self._has_next_page(),
         )
+
+        self.pages[self.page].set_page_label(current_page_label, new_page_label)
 
         if self.page_background:
             if isinstance(self.page_background, tuple):
@@ -927,7 +998,8 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
         self.text_color = tc
 
         # BEGIN Page header
-        self.header()
+        if (not self.in_toc_rendering) or self._toc_allow_page_insertion:
+            self.header()
 
         if self.line_width != lw:  # Restore line width
             self.line_width = lw
@@ -952,13 +1024,16 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
             self._write_dash_pattern(
                 dash_pattern["dash"], dash_pattern["gap"], dash_pattern["phase"]
             )
-
         # END Page header
 
     def _beginpage(
         self, orientation, format, same, duration, transition, new_page=True
     ):
         self.page += 1
+        if self.in_toc_rendering and self._toc_allow_page_insertion:
+            self._toc_inserted_pages += 1
+            self.page = len(self.pages) + 1
+            new_page = True
         if new_page:
             page = PDFPage(
                 contents=bytearray(),
@@ -1015,6 +1090,9 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
     def page_no(self):
         """Get the current page number"""
         return self.page
+
+    def get_page_label(self):
+        return self.pages[self.page].get_label()
 
     def set_draw_color(self, r, g=-1, b=-1):
         """
@@ -4727,22 +4805,25 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
 
     def _insert_table_of_contents(self):
         # Doc has been closed but we want to write to self.pages[self.page] instead of self.buffer:
-        tocp = self._toc_placeholder
+        tocp = self.toc_placeholder
         prev_page, prev_y = self.page, self.y
         self.page, self.y = tocp.start_page, tocp.y
-        # Disabling footer & header, as they have already been called:
-        self.footer = lambda *args, **kwargs: None
-        self.header = lambda *args, **kwargs: None
+        # flag rendering ToC for page breaking function
+        self.in_toc_rendering = True
         tocp.render_function(self, self._outline)
+        self.in_toc_rendering = False  # set ToC rendering flag off
         expected_final_page = tocp.start_page + tocp.pages - 1
-        if self.page != expected_final_page:
+        if self.page != expected_final_page and not self._toc_allow_page_insertion:
             too = "many" if self.page > expected_final_page else "few"
             error_msg = f"The rendering function passed to FPDF.insert_toc_placeholder triggered too {too} page breaks: "
             error_msg += f"ToC ended on page {self.page} while it was expected to span exactly {tocp.pages} pages"
             raise FPDFException(error_msg)
+        if self._toc_inserted_pages:
+            # Generating final page footer afer more pages were inserted:
+            self.in_footer = True
+            self.footer()
+            self.in_footer = False
         self.page, self.y = prev_page, prev_y
-        del self.footer
-        del self.header
 
     def file_id(self):  # pylint: disable=no-self-use
         """
@@ -5024,7 +5105,12 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
         recorder.rewind()
 
     @check_page
-    def insert_toc_placeholder(self, render_toc_function, pages=1):
+    def insert_toc_placeholder(
+        self,
+        render_toc_function: Callable,
+        pages: int = 1,
+        allow_extra_pages: bool = False,
+    ):
         """
         Configure Table Of Contents rendering at the end of the document generation,
         and reserve some vertical space right now in order to insert it.
@@ -5036,19 +5122,24 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
             pages (int): the number of pages that the Table of Contents will span,
                 including the current one that will. As many page breaks as the value of this argument
                 will occur immediately after calling this method.
+            allow_extra_pages (bool): If set to `True`, allows for an unlimited number of
+                extra pages in the ToC, which may cause discrepancies with pre-rendered
+                page numbers. For consistent numbering, using page labels to create a
+                separate numbering style for the ToC is recommended.
         """
         if not callable(render_toc_function):
             raise TypeError(
                 f"The first argument must be a callable, got: {type(render_toc_function)}"
             )
-        if self._toc_placeholder:
+        if self.toc_placeholder:
             raise FPDFException(
                 "A placeholder for the table of contents has already been defined"
-                f" on page {self._toc_placeholder.start_page}"
+                f" on page {self.toc_placeholder.start_page}"
             )
-        self._toc_placeholder = ToCPlaceholder(
+        self.toc_placeholder = ToCPlaceholder(
             render_toc_function, self.page, self.y, pages
         )
+        self._toc_allow_page_insertion = allow_extra_pages
         for _ in range(pages):
             self._perform_page_break()
 
@@ -5136,7 +5227,7 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
                 self.add_page()
             with self._marked_sequence(title=name) as struct_elem:
                 outline_struct_elem = struct_elem
-                with self._use_text_style(text_style):
+                with self.use_text_style(text_style):
                     self.multi_cell(
                         w=self.epw,
                         h=self.font_size,
@@ -5149,7 +5240,7 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
         )
 
     @contextmanager
-    def _use_text_style(self, text_style: TextStyle):
+    def use_text_style(self, text_style: TextStyle):
         if text_style:
             if text_style.t_margin:
                 self.ln(text_style.t_margin)
@@ -5279,7 +5370,7 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
             self.footer()
             self.in_footer = False
             # Generating .buffer based on .pages:
-            if self._toc_placeholder:
+            if self.toc_placeholder:
                 self._insert_table_of_contents()
             if self.str_alias_nb_pages:
                 for page in self.pages.values():