+++ fpdf/util.py
@@ -6,10 +6,12 @@ They may change at any time without prior warning or any deprecation period,
 in non-backward-compatible ways.
 """
 
-import gc, os, warnings
+import gc
+import os
+import warnings
 from numbers import Number
 from tracemalloc import get_traced_memory, is_tracing
-from typing import Iterable, Tuple, Union, NamedTuple
+from typing import Iterable, NamedTuple, Tuple, Union
 
 # default block size from src/libImaging/Storage.c:
 PIL_MEM_BLOCK_SIZE_IN_MIB = 16
@@ -136,6 +138,13 @@ def int2roman(n):
             result += numeral
             n -= integer
     return result
+
+
+def int_to_letters(n: int) -> str:
+    "Convert an integer to a letter value (A to Z for the first 26, then AA to ZZ, and so on)"
+    if n > 25:
+        return int_to_letters(int((n / 26) - 1)) + int_to_letters(n % 26)
+    return chr(n + ord("A"))
 
 
 ################################################################################